Crate X11

Reexports

pub use connection_error::ConnectionError;

pub use connection_status::ConnectionStatus;

pub use window::Window;

pub use screen_setup::ScreenSetup;

pub use window_geometry::WindowGeometry;

pub use window_children::WindowChildren;

pub use window_attribute::WindowMainAttributeSet;

pub use window_attribute::WindowSubAttributeSet;

pub use window_attribute::WindowAttributeSet;

Modules

connection_error

connection_status

input

screen_setup

window

window_attribute

window_children

window_geometry

Structs

Connection
Represents a connection to an X server. Will automatically disconnect from the X server at end of object lifetime. Guaranteed to be a valid connection upon successful construction but not after.

Coordinate
A 2D coordinate where (0, 0) is the upper left corner.

RectangularSize
A simple representation for the size of a rectangle.

RequestDelay
Represents a request that is waiting to be sent. Flushes all pending requests upon destruction.

RequestError

Screen

Setup
See: http://xcb.freedesktop.org/manual/group__XCB__Core__API.html#gafc379a27800bf735818a0760bd036321

Traits

Cookie
Implementors of this trait represent a pending reply from the X server made via an asynchronous request.

Reply