Module X11​::Window_Attribute​

Reexports

pub use self::Back_Pixmap::BackPixmapSet;

pub use self::Bit_Gravity::BitGravity;

pub use self::Win_Gravity::WinGravity;

pub use self::Backing_Store::BackingStore;

pub use self::Event::EventSet;

pub use self::Colormap::ColormapSet;

pub use self::Cursor::CursorSet;

Modules

Back_Pixmap

Backing_Store

Bit_Gravity

Colormap

Cursor

Event

Win_Gravity

Structs

WindowAttributeSet
A struct which can be modeled as an algebraic datatype consisting of a unique data constructor for each element of the 15-ary Cartesian product of back_pixmap × back_pixel × ... × cursor where back_pixmap, back_pixel, ... cursor are modeled as sets each equal to { “on”, “off” }. (i.e. data WindowAttributeSet = back_pixmap_on__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_off__...__cursor_on ... -- Obviously listing 2¹⁵ data constructors is -- impractical. ) Indicate the data constructor one intends to use via the main_attributes field. Additionally, some of the data constructors (specifically the ones indicating that back_pixmap, bit_gravity, win_gravity, backing_store, event, colormap, or cursor are “on”) take arguments. Indicate the arguments to the data constructor one intends to use via the sub_attributes field.

WindowMainAttributeSet

WindowSubAttributeSet
See documentation for WindowAttributeSet.

Statics

pub static back_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static back_pixmap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_planes: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_store: WindowMainAttributeSet = BitFlags { bits: $value }

pub static bit_gravity: WindowMainAttributeSet = BitFlags { bits: $value }

pub static border_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static border_pixmap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static colormap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static cursor: WindowMainAttributeSet = BitFlags { bits: $value }

pub static dont_propagate: WindowMainAttributeSet = BitFlags { bits: $value }

pub static event: WindowMainAttributeSet = BitFlags { bits: $value }

pub static override_reddirect: WindowMainAttributeSet = BitFlags { bits: $value }

pub static save_under: WindowMainAttributeSet = BitFlags { bits: $value }

pub static win_gravity: WindowMainAttributeSet = BitFlags { bits: $value }

Type Definitions

WindowMainAttributeInt