Struct X11​::Window_Attribute​::WindowAttributeSet​

pub struct WindowAttributeSet {
    pub main_attributes: WindowMainAttributeSet,
    pub sub_attributes: WindowSubAttributeSet,
}

A struct which can be modeled as an algebraic datatype consisting of a unique data constructor for each element of the 15-ary Cartesian product of back_pixmap × back_pixel × ... × cursor where back_pixmap, back_pixel, ... and cursor are modeled as sets each equal to { “on”, “off” }. (i.e. data WindowAttributeSet = back_pixmap_on__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_off__...__cursor_on ... -- Obviously listing 2¹⁵ data constructors is impractical. ) Indicate the data constructor one intends to use via the main_attributes field. Additionally, some of the data constructors (specifically the ones indicating that back_pixmap, bit_gravity, win_gravity, backing_store, event, colormap, or cursor are “on”) take arguments. Indicate the arguments to the data constructor one intends to use via the sub_attributes field.

Fields

main_attributes

sub_attributes

Methods

impl WindowAttributeSet

fn new(main_attributes: WindowMainAttributeSet, sub_attributes: WindowSubAttributeSet) -> WindowAttributeSet

fn main_attributes(&self) -> WindowMainAttributeSet

fn sub_attributes(&self) -> WindowSubAttributeSet

fn sub_attribute_array(&self) -> [u32, ..7]