Module X11​::Window_Geometry​

Structs

Cookie

Reply
The reply from the X server holding the requested window's geometetrical information.

WindowGeometry
Holds the position, size, and border width of a window.

Functions

make_request