Enum X11::connection_error::ConnectionError

pub enum ConnectionError {
    Generic,
    ExtNotSupported,
    MemInsufficient,
    ReqLenExceeded,
    ParseErr,
    InvalidScreen,
    FDPassingFailure,
}

Variants

Generic

Socket error, pipe error, or other stream error

ExtNotSupported

Extension not supported

MemInsufficient

Memory not available

ReqLenExceeded

Exceeding request length for server

ParseErr

Unable to parse display string

InvalidScreen

No screen matching display on server (The display is usually specified with the $DISPLAY environment variable.)

FDPassingFailure

File descriptor passing operation failure (This is not explicitly stated as a possible return value in the comments above the declaration of xcb_connection_has_error in xcb.h, but it is implied as a possible return value from the macro definition XCB_CONN_CLOSED_FDPASSING_FAILED.)