Struct X11​::RequestDelay​

pub struct RequestDelay<'a> {
    // some fields omitted
}

Represents a request that is waiting to be sent. Flushes all pending requests upon destruction.

Methods

impl<'a> RequestDelay<'a>

fn new(connection: &'a Connection) -> RequestDelay<'a>

fn force(self)

Use force to flush all pending requests. The RequestDelay called with the force method is moved into the force method where its destructor is called (once).

fn subsume(&self, other: RequestDelay)

Use subsume to prevent other RequestDelays in the current scope from calling their destructors (and flushing pending requests). The RequestDelay placed in the “other” parameter is moved into the subsume method where its destructor is made to do nothing. No error can occur if subsume is not used, but it helps control exactly when pending requests are flushed.