Trait X11​::Cookie​

pub trait Cookie<'a, R: Reply>: Drop {
    fn wait_for_reply(self) -> Result<R, RequestError>;
}

Implementors of this trait represent a pending reply from the X server made via an asynchronous request.

Required Methods

fn wait_for_reply(self) -> Result<R, RequestError>