Module X11::window_attribute

Reexports

pub use self::background_pixmap::BackgroundPixmap;

pub use self::bit_gravity::BitGravity;

pub use self::win_gravity::WinGravity;

pub use self::backing_store::BackingStore;

pub use self::event::EventSet;

pub use self::colormap::ColormapSet;

pub use self::cursor::CursorSet;

Modules

background_pixmap

backing_store

bit_gravity

border_pixmap

colormap

cursor

event

win_gravity

Structs

WindowAttributeSet
A struct which can be modeled as an algebraic datatype consisting of a unique data constructor for each element of the 15-ary Cartesian product of back_pixmap × back_pixel × ... × cursor where back_pixmap, back_pixel, ... and cursor are modeled as sets each equal to { “on”, “off” }. (i.e. data WindowAttributeSet = back_pixmap_on__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_on__...__cursor_on | back_pixmap_off__back_pixel_off__...__cursor_on ... -- Obviously listing 2¹⁵ data constructors is impractical. ) Indicate the data constructor one intends to use via the main_attributes field. Additionally, some of the data constructors (specifically the ones indicating that back_pixmap, bit_gravity, win_gravity, backing_store, event, colormap, or cursor are “on”) take arguments. Indicate the arguments to the data constructor one intends to use via the sub_attributes field.

WindowMainAttributeSet

WindowSubAttributeSet
See documentation for WindowAttributeSet.

Statics

pub static background_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static background_pixmap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_planes: WindowMainAttributeSet = BitFlags { bits: $value }

pub static backing_store: WindowMainAttributeSet = BitFlags { bits: $value }

pub static bit_gravity: WindowMainAttributeSet = BitFlags { bits: $value }

pub static border_pixel: WindowMainAttributeSet = BitFlags { bits: $value }

pub static border_pixmap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static colormap: WindowMainAttributeSet = BitFlags { bits: $value }

pub static cursor: WindowMainAttributeSet = BitFlags { bits: $value }

pub static dont_propagate: WindowMainAttributeSet = BitFlags { bits: $value }

pub static event: WindowMainAttributeSet = BitFlags { bits: $value }

pub static override_reddirect: WindowMainAttributeSet = BitFlags { bits: $value }

pub static save_under: WindowMainAttributeSet = BitFlags { bits: $value }

pub static win_gravity: WindowMainAttributeSet = BitFlags { bits: $value }

Type Definitions

WindowMainAttributeInt
A bitmask of 15 flags where the flags are back_pixmap, back_pixel, ... cursor.