Struct X11​::Connection​

pub struct Connection {
    // some fields omitted
}

Represents a connection to an X server. Will automatically disconnect from the X server at end of object lifetime. Guaranteed to be a valid connection upon successful construction but not after.

Methods

impl Connection

fn new() -> Result<Connection, ConnectionError>

Construct a connection to the X server only if it can be shown to be an initially valid connection. Connects to the X server specified by the $DISPLAY environment variable (if $DISPLAY can be parsed).

fn new_with_default_screen() -> (Result<Connection, ConnectionError>, Option<Screen>)

Does the same as new() but also returns the default Screen if one exists.

fn status(&self) -> ConnectionStatus

Test if connected to the X server and if not return why.

fn is_valid(&self) -> bool

Test if connected to the X server.

fn flush(&self)

fn setup<'a>(&'a self) -> Setup<'a>

See: http://xcb.freedesktop.org/manual/group__XCB__Core__API.html#gafc379a27800bf735818a0760bd036321

fn make_window_geometry_request<'a>(&'a self, window: Window) -> (Cookie<'a>, RequestDelay<'a>)

fn make_window_children_request<'a>(&'a self, window: Window) -> (Cookie<'a>, RequestDelay<'a>)

fn change_window_attributes<'a>(&'a self, window: Window, value_mask: WindowAttributeSet, values: &[u32]) -> RequestDelay<'a>